% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allen.R
\name{allen_context}
\alias{allen_context}
\title{Generate LLM-friendly context for the brain.map.org package}
\usage{
allen_context()
}
\value{
Character string (invisibly), also printed
}
\description{
Generate LLM-friendly context for the brain.map.org package
}
